requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-braces==1.13.0
django-cors-middleware==1.3.1
django-daterange-filter==1.3.0
django-excel==0.0.10
django-import-export==1.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
lml==0.0.9
mysqlclient==1.4.1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pytz==2018.9
PyYAML==5.1
requests==2.21.0
simplejson==3.16.0
tablib==0.13.0
texttable==1.6.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
